Updated System

+++ admin/user/controller.php
@@ -94,15 +94,26 @@ switch ($action) {
 
 		// 	$user = New User();
 		// 	$user->delete($id[$i]);
-
+		//$user = $_SESSION['ADMIN_USERID'];
 		
 				$id = 	$_GET['id'];
+				$mydb = new Database();
+				$sql = "SELECT * FROM tblusers WHERE USERID = $id";
+				$mydb->setQuery($sql);
+ 				$res = $mydb->loadSingleResult();
 
-				$user = New User();
-	 		 	$user->delete($id);
-			 
-			message("User has been deleted!","info");
-			redirect('index.php');
+				if ($_SESSION['ADMIN_ROLE'] != 'Administrator'){
+					message("Only Administrator Can Delete!","info");
+				}else{
+					if($res->DELETEABLE == true){
+						$user = New User();
+						$user->delete($id);
+						message("User has been deleted!","info");
+						
+					}
+				}
+				redirect('index.php');
+			
 		// }
 		// }
